backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.7

# HTML解析
beautifulsoup4==4.12.2
lxml==4.9.3

# 数据分析
pandas==2.1.3
numpy==1.25.2

# 测试
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# 开发工具
black==23.11.0
flake8==6.1.0
isort==5.12.0

# P2-LD-014: Celery异步任务集成
celery==5.3.4
redis==5.0.1
kombu==5.3.4
billiard==4.2.0
vine==5.1.0

# 监控和调试（可选）
flower==2.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.0

# 其他实用工具
python-decouple==3.8
Pillow==10.1.0
gunicorn==21.2.0
django-extensions==3.2.3
